Updated Shooting Mechanics
Enemies and Player will not have bullets that aren't parented to them. Currently it's a little awkward since it only spawns on the front of the player and the left of enemies. Will do some more work, but this set up should work for some basic left to right running gameplay